refactor(battlepass): Refactors battlepass system and fixes clan economy types
Changes clan coins from float to int throughout the codebase for consistency and precision in currency handling.

Adds battlepass reward configuration models and integrates battlepass levels into guild economy config with JSON storage.

Updates battlepass default level from 0 to 1 and adjusts point rewards (8 points per message, 100 points on level up).

Fixes clan shop and improvements autocomplete to only send item identifiers instead of embedded prices, preventing parsing issues and improving validation by checking prices against guild config.

Removes unused FAQ view initialization and cleans up Dockerfile comments.

Adds TYPE_CHECKING imports to reduce circular dependency issues.

+++ src/infra/db/models/clan.py
@@ -29,7 +29,7 @@ class Clan(IdIntegerMixin, Base, CreatedAtMixin):
     )  # Discord role for the clan
 
     # Economy & progression
-    coins: Mapped[float] = mapped_column(nullable=False, default=0.0)
+    coins: Mapped[int] = mapped_column(nullable=False, default=0)
     current_exp: Mapped[int] = mapped_column(
         Integer, nullable=False, default=0
     )
